requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=1.3.0=pyhd8ed1ab_0
aiohttp=3.8.3=py39ha30fb19_1
aiosignal=1.3.1=pyhd8ed1ab_0
appnope=0.1.3=pyhd8ed1ab_0
asttokens=2.1.0=pyhd8ed1ab_0
async-timeout=4.0.2=py39hecd8cb5_0
attrs=22.1.0=pyh71513ae_1
autopep8=1.6.0=pyhd3eb1b0_1
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
blinker=1.5=pyhd8ed1ab_0
boost=1.78.0=py39h6dc771e_4
boost-cpp=1.78.0=h8b082ac_1
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
brotlipy=0.7.0=py39h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
c-ares=1.18.1=h0d85af4_0
ca-certificates=2022.9.24=h033912b_0
cachetools=5.2.0=pyhd8ed1ab_0
cairo=1.16.0=h691a603_2
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py39hc55c11b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=unix_pyhd8ed1ab_2
comm=0.1.0=pyhd8ed1ab_0
cryptography=38.0.1=py39hf6deb26_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.3=py39h7a8716b_1
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
expat=2.5.0=hf0c8a7f_0
ffmpeg=4.3=h0a44026_0
fftw=3.3.9=h9ed2024_1
fontconfig=2.14.1=h5bb23bf_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=hd8bbffd_0
frozenlist=1.3.3=py39ha30fb19_0
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
glib=2.69.1=h8346a28_1
gmp=6.2.1=he9d5cce_3
gnutls=3.6.15=hed9c0bf_0
google-auth=2.14.1=pyh1a96a4e_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
greenlet=2.0.1=py39h7a8716b_0
grpcio=1.49.1=py39h69d5c6f_1
icu=70.1=h96cf925_0
idna=3.4=py39hecd8cb5_0
importlib-metadata=5.1.0=pyha770c72_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.18.0=pyh736e0ef_0
ipython=8.6.0=pyhd1c38e8_1
jedi=0.18.1=pyhd8ed1ab_2
jinja2=3.1.2=py39hecd8cb5_0
joblib=1.1.1=py39hecd8cb5_0
jpeg=9e=hca72f7f_0
jupyter_client=7.4.7=pyhd8ed1ab_0
jupyter_core=5.0.0=py39h6e9494a_0
kiwisolver=1.4.2=py39he9d5cce_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libabseil=20220623.0=cxx17_h844d122_5
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libcxx=14.0.6=h9765a3e_0
libdeflate=1.8=h9ed2024_5
libffi=3.3=hb1e8313_2
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libgrpc=1.49.1=h834a566_1
libiconv=1.16=hca72f7f_2
libidn2=2.3.2=h9ed2024_0
libpng=1.6.37=ha441bb4_0
libprotobuf=3.21.9=hbc0c0cd_0
libsodium=1.0.18=hbcb3906_1
libtasn1=4.16.0=h9ed2024_0
libtiff=4.4.0=h2cd0358_2
libunistring=0.9.10=h9ed2024_0
libwebp=1.2.4=h56c3ce4_0
libwebp-base=1.2.4=hca72f7f_0
libxml2=2.9.14=hea49891_4
libzlib=1.2.13=hfd90126_4
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.3=h23ab428_1
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.1=py39hca72f7f_0
matplotlib=3.5.3=py39hecd8cb5_0
matplotlib-base=3.5.3=py39hfb0c5b7_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
multidict=6.0.2=py39ha30fb19_2
munkres=1.1.4=py_0
ncurses=6.3=hca72f7f_3
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=h230ac6f_1
networkx=2.8.4=py39hecd8cb5_0
numpy=1.23.4=py39he696674_0
numpy-base=1.23.4=py39h9cd3388_0
oauthlib=3.2.2=pyhd8ed1ab_0
openh264=2.1.1=h8346a28_0
openssl=1.1.1s=hfd90126_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.2=py39hecff1ad_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=he49afe7_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.2.0=py39hde71d04_1
pip=22.2.2=py39hecd8cb5_0
pixman=0.40.0=hbcb3906_0
platformdirs=2.5.2=pyhd8ed1ab_1
prompt-toolkit=3.0.33=pyha770c72_0
protobuf=4.21.9=py39h7a8716b_0
psutil=5.9.4=py39ha30fb19_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycairo=1.22.0=py39h41776c8_0
pycodestyle=2.8.0=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pyg=2.1.0=py39_torch_1.13.0_cpu
pygments=2.13.0=pyhd8ed1ab_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39hecd8cb5_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.15=hdfd78df_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.13.0=py3.9_0
pytorch-cluster=1.6.0=py39_torch_1.13.0_cpu
pytorch-scatter=2.0.9=py39_torch_1.13.0_cpu
pytorch-sparse=0.6.15=py39_torch_1.13.0_cpu
pytz=2022.6=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyzmq=24.0.1=py39hed8f129_1
rdkit=2022.09.1=py39h693e27b_1
re2=2022.06.01=hb486fe8_1
readline=8.2=hca72f7f_0
reportlab=3.5.68=py39hf37cc50_1
requests=2.28.1=py39hecd8cb5_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.1.3=py39he9d5cce_0
scipy=1.9.3=py39h3d31255_0
setuptools=65.5.0=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlalchemy=1.4.44=py39ha30fb19_0
sqlite=3.39.3=h707629a_0
stack_data=0.6.1=pyhd8ed1ab_0
tensorboard=2.11.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py39h7eb6a14_3
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pyhd3eb1b0_0
torch-tb-profiler=0.4.0=pypi_0
torchvision=0.14.0=py39_cpu
tornado=6.2=py39hca72f7f_0
tqdm=4.64.1=py39hecd8cb5_0
traitlets=5.5.0=pyhd8ed1ab_0
typing_extensions=4.3.0=py39hecd8cb5_0
tzdata=2022f=h04d1e81_0
urllib3=1.26.12=py39hecd8cb5_0
wcwidth=0.2.5=pyh9f0ad1d_2
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=hca72f7f_0
yarl=1.8.1=py39ha30fb19_0
zeromq=4.3.4=he49afe7_1
zipp=3.10.0=pyhd8ed1ab_0
zlib=1.2.13=hfd90126_4
zstd=1.5.2=hcb37349_0
